experimenting with typescript defuns
