+++ modalities/dom/components/xen/xen-state.d.ts
@@ -1,0 +1,3 @@
+//declare module "xen-state";
+
+export function XenStateMixin(Base: any);
